+++ gulpfile.js
@@ -1,8 +1,11 @@
-var gulp = require('gulp');
-var rename = require("gulp-rename");
-var cleanCSS = require('gulp-clean-css');
-var uglify = require('gulp-uglify');
-var concat = require('gulp-concat');
+var gulp        = require('gulp');
+var rename      = require("gulp-rename");
+var cleanCSS    = require('gulp-clean-css');
+var uglify      = require('gulp-uglify');
+var concat      = require('gulp-concat');
+var bs = require('browser-sync').create();
+var nodemon = require('gulp-nodemon');
+
 
 // Directories
 var source = './src';
@@ -57,13 +60,43 @@ gulp.task('vendorjs', ()=> {
       .pipe(gulp.dest(destination + '/js'));
 });
 
+gulp.task('browser-sync', ['nodemon'], () => {
+    bs.init(null, {
+      proxy: "http://localhost:8080",
+        files: [{
+                match: [source + '/**/*.*'],
+                fn:    function (event, file) {
+                    this.reload()
+                }
+            }],
+        port: 3000
+    });
+});
+
+gulp.task('nodemon', function (cb) {
+
+	var started = false;
+
+	return nodemon({
+		script: 'app.js'
+	}).on('start', function () {
+		if (!started) {
+			cb();
+			started = true;
+		}
+	});
+});
+
 // Build task
 gulp.task('build', ['css', 'img', 'js', 'vendorcss', 'vendorjs', 'sounds', 'fonts']);
 
-// Default task : build
-gulp.task('default', ['build']);
-
-// Watch task
-gulp.task('watch', function () {
+// Watch for changes and build
+gulp.task('watch', () => {
   gulp.watch(source + '/**/*.*', ['build']);
 });
+
+// Serve: open the browser on port 3000, watch for changes and reload browser
+gulp.task('serve', ['watch', 'browser-sync']);
+
+// Default task: serve
+gulp.task('default', ['serve']);