+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+Version 1.2.0 (11/06/2018)
+-----------------------------
+* New Gulp task: 'gulp serve' (open the app in the browser on http://localhost:3000, watch for changes in /src folder and reload the browser)
+* New functionality: shake screen when a sound is played
+
 Version 1.1.0 (10/06/2018)
 -----------------------------
 * Upgrade socket.io from version 2.0.4 to 2.1.1